Major refactoring, moving REST client code to sub-directory of http/rexos
