+++ http/rexos/gorex.go
@@ -7,14 +7,3 @@
 // account.  Once you activated your account, you can simply create an API access token with a
 // `ClientId` and a `ClientSecret`.
 package rexos
-
-import (
-	"strings"
-)
-
-// StripTemplateParameter removes the trailing template parameters of an HATEOAS URL
-//
-// For example: "https://rex.robotic-eyes.com/rex-gateway/api/v2/rexReferences/1000/project{?projection}"
-func StripTemplateParameter(templateURL string) string {
-	return strings.Split(templateURL, "{")[0]
-}
