+++ http/rexos/rest/client_test.go
@@ -1,4 +1,4 @@
-package rexos
+package rest
 
 import (
 	"testing"
