+++ http/rexos/rest/strip_test.go
@@ -1,4 +1,4 @@
-package rexos
+package rest
 
 import (
 	"testing"
